+++ Cartoonify-Image/Cartoonify using KMeans/Script/vid.py
