+++ Cartoonify-Image/Cartoonify-GUI/Readme.md
@@ -1,0 +1,50 @@
+## Cartoonify Image and save it
+### This is a very baisc GUI for cartoonifying iage and save it
+## How to start
+
+
+- Fork the repository
+
+- Git clone your forked repository
+- Create virtual environment-
+```
+- python -m venv env
+- source env/bin/activate (Linux)
+- env\Scripts\activate (Windows)
+```
+- Install dependencies
+- Go to project directory
+```
+- cd Cartoonify Image
+```
+- install these requirements
+
+```
+-  pip install opencv-python
+- pip install tkinter
+- pip install easygui
+- pip install pillow
+```  
+- Open Terminal
+```
+python cartoonify_GUI.py
+```
+
+### Video recording of the process
+
+[![YouTube link](mq2.jpg)](https://youtu.be/VDqEv6_FDt4 "Cartoonify GUI")
+
+
+### Screenshots: 
+<img src="mq2.jpg" height="300px">
+
+### GUI Interface
+<img src="Updated GUI.png" height="600px">
+
+### Cartoonified Image 
+<img src="cartoonified.jpg" height="600px">
+
+### Original Image
+<img src="wp2030093.jpg" height="600px">
+
+
