+++ Cartoonify-Image/Cartoonify-GUI/Updated GUI.png
