+++ Cartoonify-Image/Cartoonized-Image/readme.md
@@ -1,0 +1,3 @@
+This application uses openCV to create cartoonish filter for any image with the help of Gaussian Filter
+
+Gaussian Filter- Uses a blurring function, adaptive thresholding and dilation to expose the main features of an image.
